requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2